added balance showing feature on multiple screens

+++ app/src/main/java/com/example/tbilisi_parking_final_exm/presentation/screen/user_panel/wallet/balance/BalanceFragment.kt
@@ -92,7 +92,7 @@ class BalanceFragment : BaseFragment<FragmentBalanceBinding>(FragmentBalanceBind
                 cvv = etCVV,
                 cardId = args.cardId?.toIntOrNull(),
                 isRememberCardChecked = chkRememberCard.isChecked,
-                amount = args.amount
+                amount = args.amount.toDouble()
             )
         )
     }
